# Fix unchecked vector pre-allocation