--- a/rmpv/src/decode/value.rs
+++ b/rmpv/src/decode/value.rs
@@ -1,3 +1,4 @@
+use std::cmp::min;
 use std::io::{self, Read};
 
 use rmp::Marker;
@@ -7,6 +8,11 @@ use rmp::decode::{read_marker, read_data_u8, read_data_u16, read_data_u32, read_
 
 use crate::{Utf8String, Value};
 use super::Error;
+
+
+// See https://github.com/3Hren/msgpack-rust/issues/151
+const PREALLOC_MAX: usize = 64 * 1024; // 64 KiB
+
 
 fn read_array_data<R: Read>(rd: &mut R, mut len: usize) -> Result<Vec<Value>, Error> {
     // Note: Do not preallocate a Vec of size `len`.
@@ -48,9 +54,7 @@ fn read_str_data<R: Read>(rd: &mut R, len: usize) -> Result<Utf8String, Error> {
 }
 
 fn read_bin_data<R: Read>(rd: &mut R, len: usize) -> Result<Vec<u8>, Error> {
-    // Note: Do not preallocate a Vec of size `len`.
-    // See https://github.com/3Hren/msgpack-rust/issues/151
-    let mut buf = Vec::new();
+    let mut buf = Vec::with_capacity(min(len, PREALLOC_MAX));
     let bytes_read = rd.take(len as u64).read_to_end(&mut buf).map_err(Error::InvalidDataRead)?;
     if bytes_read != len {
         return Err(Error::InvalidDataRead(io::Error::new(