@kornelski 2019-10-04
I'm concerned that this may be too drastic. Would it suffice to put some arbitrary limit on it? e.g. `with_capacity(len.min(LIMIT))` where the limit could be something like 4K or 64K?

This would reduce the problem from outright DoS to mere amplification of memory usage in the worst case of infinitely recursive arrays overpromising their size. Amplification seems like a lesser problem, and also likely happens already anyway (all memory management has some overhead, even non-preallocated vec will grow by doubling its size, so there's always 2x amplification there).