@dbrgn 2019-10-04
And in any case, we are speculating and should probably have some benchmarks.

@kornelski the current rmpv benchmarks are from 2017 and broken. Do we still need them, or can they be disabled or removed for now?