@kornelski 2019-10-04
It'd be nice to fix the benchmarks, so we don't have to speculate :)

Reserving capacity with a limit for non-nested types, and no reservation for nested types, sounds safe and fast enough.