@dbrgn 2019-10-08
When preallocating up to 64 KiB for bin data:

```
test read_bin32_100kib                    ... bench:       2,347 ns/iter (+/- 403) = 43630 MB/s
test read_bin32_100mib                    ... bench:  48,031,872 ns/iter (+/- 1,750,672) = 2183 MB/s
test read_bin32_1mib                      ... bench:      33,701 ns/iter (+/- 719) = 31114 MB/s
test read_bin32_20mib                     ... bench:   9,528,805 ns/iter (+/- 729,030) = 2200 MB/s
test read_bin32_50kib                     ... bench:       1,102 ns/iter (+/- 50) = 46460 MB/s
```

Comparison:

```
                 prealloc         no-prealloc      prelloc max64k
bin32  50 kib       1,887       1,393 (- 26%)       1,102 (- 42%)
bin32 100 kib       3,845       2,584 (- 33%)       2,347 (- 39%)
bin32   1 mib      54,690      33,507 (- 39%)      33,701 (- 38%)
bin32  20 mib   2,587,674  10,209,202 (+294%)   9,528,805 (+268%)
bin32 100 mib  47,660,737  47,599,682 (-  0%)  48,031,872 (-  1%)
```

Doesn't hurt, but doesn't help a lot either.