@dbrgn 2019-10-08
Rebased against master.